Moved tank and projectile positioning to float

+++ server/configuration.c
@@ -16,7 +16,7 @@ struct configuration* configuration_alloc(){
     return self;
 }
 
-void configuration_set_values(struct configuration* self, uint32_t width, uint32_t height, uint32_t background_scale, uint32_t players_count, uint32_t player_id, uint32_t tank_spawn_x,uint32_t tank_spawn_y, uint32_t map_number){
+void configuration_set_values(struct configuration* self, uint32_t width, uint32_t height, uint32_t background_scale, uint32_t players_count, uint32_t player_id, float tank_spawn_x, float tank_spawn_y, uint32_t map_number){
     self->width = width;
     self->height = height;
     self->background_scale = background_scale;
