+++ server/tank.c
@@ -17,7 +17,7 @@ struct tank* tank_alloc(){
     return self;
 }
 
-void tank_set_values(struct tank* self, uint32_t player_id, uint32_t x, uint32_t y, float tank_angle, float hp, float turret_angle, uint32_t tank_version){
+void tank_set_values(struct tank* self, uint32_t player_id, float x, float y, float tank_angle, float hp, float turret_angle, uint32_t tank_version){
     self->player_id = player_id;
     self->x = x;
     self->y = y;
@@ -27,7 +27,7 @@ void tank_set_values(struct tank* self, uint32_t player_id, uint32_t x, uint32_t
     self->tank_version = tank_version;
 }
 
-void tank_update(struct tank* self, uint32_t x, uint32_t y, float tank_angle, float hp, float turret_angle){
+void tank_update(struct tank* self, float x, float y, float tank_angle, float hp, float turret_angle){
     self->x = x;
     self->y = y;
     self->tank_angle = tank_angle;
